// Application.RMGUI/ClientApp/src/app/models/ApplicantProfileReview.ts
export class ApplicantProfileReview {
  applicant_profile_review_id: string;
  panel_name: string;
  applicant_name: string;
  applicant_profile_status: string;
  panel_profile_review: string;
  
}
//applicant_profile_review_id
//panel_name
//applicant_name
//applicant_profile_status
//panel_profile_review